code refactoring : move class to subpackage
