--- bundles/org.openhab.binding.linky/src/main/java/org/openhab/binding/linky/internal/ValueType.java
@@ -1,28 +1,0 @@
-/*
- * Copyright (c) 2010-2025 Contributors to the openHAB project
- *
- * See the NOTICE file(s) distributed with this work for additional
- * information.
- *
- * This program and the accompanying materials are made available under the
- * terms of the Eclipse Public License 2.0 which is available at
- * http://www.eclipse.org/legal/epl-2.0
- *
- * SPDX-License-Identifier: EPL-2.0
- */
-package org.openhab.binding.linky.internal;
-
-import org.eclipse.jdt.annotation.NonNullByDefault;
-
-/**
- * Defines all the type of values
- *
- * @author Olivier MARCEAU - Initial contribution
- *
- */
-@NonNullByDefault
-public enum ValueType {
-    INTEGER,
-    STRING,
-    DATE
-}
