+++ bundles/org.openhab.binding.linky/src/main/java/org/openhab/binding/linky/internal/factory/LinkyHandlerFactory.java
@@ -10,10 +10,10 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  */
-package org.openhab.binding.linky.internal;
+package org.openhab.binding.linky.internal.factory;
 
 import static java.time.temporal.ChronoField.*;
-import static org.openhab.binding.linky.internal.LinkyBindingConstants.*;
+import static org.openhab.binding.linky.internal.constants.LinkyBindingConstants.*;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
