+++ bundles/org.openhab.binding.linky/src/main/java/org/openhab/binding/linky/internal/handler/ThingTariffHpHcHandler.java
@@ -12,7 +12,7 @@
  */
 package org.openhab.binding.linky.internal.handler;
 
-import static org.openhab.binding.linky.internal.LinkyBindingConstants.*;
+import static org.openhab.binding.linky.internal.constants.LinkyBindingConstants.*;
 
 import java.util.List;
 import java.util.stream.Stream;
