+++ bundles/org.openhab.binding.linky/src/main/java/org/openhab/binding/linky/internal/helpers/LinkyFrame.java
@@ -10,7 +10,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  */
-package org.openhab.binding.linky.internal;
+package org.openhab.binding.linky.internal.helpers;
 
 import java.io.Serializable;
 import java.util.EnumMap;
@@ -20,6 +20,8 @@ import org.eclipse.jdt.annotation.NonNullByDefault;
 import org.eclipse.jdt.annotation.Nullable;
 import org.openhab.binding.linky.internal.types.Evolution;
 import org.openhab.binding.linky.internal.types.FrameType;
+import org.openhab.binding.linky.internal.types.InvalidFrameException;
+import org.openhab.binding.linky.internal.types.LinkyChannel;
 import org.openhab.binding.linky.internal.types.Phase;
 import org.openhab.binding.linky.internal.types.Pricing;
 import org.openhab.binding.linky.internal.types.TeleinfoTicMode;
