+++ bundles/org.openhab.binding.linky/src/main/java/org/openhab/binding/linky/internal/helpers/LinkyFrameUtil.java
@@ -10,7 +10,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  */
-package org.openhab.binding.linky.internal;
+package org.openhab.binding.linky.internal.helpers;
 
 import org.eclipse.jdt.annotation.NonNullByDefault;
 import org.openhab.binding.linky.internal.types.TeleinfoTicMode;
