+++ bundles/org.openhab.binding.linky/src/main/java/org/openhab/binding/linky/internal/types/FrameType.java
@@ -12,7 +12,7 @@
  */
 package org.openhab.binding.linky.internal.types;
 
-import static org.openhab.binding.linky.internal.LinkyBindingConstants.*;
+import static org.openhab.binding.linky.internal.constants.LinkyBindingConstants.*;
 
 import org.eclipse.jdt.annotation.NonNullByDefault;
 import org.eclipse.jdt.annotation.Nullable;
