+++ bundles/org.openhab.binding.linky/src/main/java/org/openhab/binding/linky/internal/types/ValueType.java
@@ -23,5 +23,6 @@ import org.eclipse.jdt.annotation.NonNullByDefault;
 @NonNullByDefault
 public enum ValueType {
     INTEGER,
-    STRING
+    STRING,
+    DATE
 }
